+++ tests/cosmic/mod.rs
@@ -1,3 +1,3 @@
 mod bplane;
 mod eclipse;
-// mod orbit; // TODO(ANISE): Remove this after confirmation that it was all moved to ANISE
+mod orbit_dual;
